strdup

+++ 0x0B-malloc_free/1-strdup.c
@@ -18,7 +18,7 @@ char *_strdup(char *str)
 	{
 		return (NULL);
 	}
-	for (i = 0; i < strlen(str); i++)
+	for (i = 0; i < (int)strlen(str); i++)
 	{
 		dupstr[i] = str[i];
 	}
